requirements.txt
django
djangorestframework
markdown
django-filter
mysqlclient
pyyaml
requests
